+++ src/main/java/de/samply/frontend/dto/DtoFactory.java
@@ -1,9 +1,9 @@
 package de.samply.frontend.dto;
 
-import de.samply.project.state.ProjectBridgeheadState;
+import de.samply.db.model.NotificationUserAction;
 import jakarta.validation.constraints.NotNull;
 
-import java.time.Instant;
+import java.util.function.Supplier;
 
 public class DtoFactory {
 
@@ -29,7 +29,7 @@ public class DtoFactory {
         );
     }
 
-    public static Notification convert(@NotNull de.samply.db.model.Notification notification) {
+    public static Notification convert(@NotNull de.samply.db.model.Notification notification, Supplier<NotificationUserAction> userActionSupplier) {
         return new Notification(
                 notification.getEmail(),
                 notification.getTimestamp(),
@@ -37,7 +37,9 @@ public class DtoFactory {
                 notification.getBridgehead(),
                 notification.getOperationType(),
                 notification.getDetails(),
-                notification.getError()
+                notification.getError(),
+                notification.getHttpStatus(),
+                userActionSupplier.get().isRead()
         );
     }