+++ src/main/java/de/samply/frontend/dto/Notification.java
@@ -1,6 +1,7 @@
 package de.samply.frontend.dto;
 
 import de.samply.notification.OperationType;
+import org.springframework.http.HttpStatus;
 
 import java.time.Instant;
 
@@ -11,6 +12,8 @@ public record Notification(
         String bridgehead,
         OperationType operationType,
         String details,
-        String error
+        String error,
+        HttpStatus httpStatus,
+        Boolean read
 ) {
 }